project-b/pyproject.toml
[build-system]
requires = ["setuptools >= 68", "wheel","setuptools-git >= 1.1", "pytest-runner", "toml", "reahl-component-metadata >= 7.0.0"]
build-backend = "setuptools.build_meta"

  
[project]
name = "project-b"
version = "3.1"
dependencies = [
  "project-c>=4.0,<4.1",
  "project-a>=1.4,<1.5",
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]

[tool.setuptools.packages.find]
exclude = ["etc", "build", "dist"]

[tool.reahl-component]
persisted = [
  "projectb.code:BObject"
]

[tool.reahl-component.versions.'3.1']
migrations = [
  "projectb.migrations:ToNewC40"
]

[tool.reahl-component.versions.'3.0']
dependencies = [
  "project-c>=3.1,<3.2",
  "project-a>=1.3,<1.4",
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]
migrations = [
  "projectb.migrations:ToNewC31"
]


[tool.reahl-component.versions.'2.0']
dependencies = [
  "project-c>=3.0,<3.1",
  "reahl-component>=7.0,<7.1",
  "reahl-sqlalchemysupport>=7.0,<7.1"
]
migrations = [
  "projectb.migrations:CreateDB"
]
